Added the interop module (breaking)
