+++ Cargo.toml
@@ -24,9 +24,3 @@ maintenance = { status = "passively-maintained" }
 [dependencies]
 openvr_sys = "2.0.3"
 lazy_static = "1.3.0"
-vk-sys = {version = "0.5", optional = true}
-
-[features]
-default = []
-vulkan = ["vk-sys"]  # Enables Vulkan interface functions.
-#d3d12 = []   # Enables D3D12 interface functions.
