+++ src/interop/mod.rs
@@ -1,0 +1,4 @@
+//! Types required for interoperating with graphics APIs supported by OpenVR.
+//!
+//! See the individual submodules depending on the API you're using.
+pub mod vulkan;
